feat: add basic CI (#3)
### Summary  
This PR adds a basic end-to-end (e2e) workflow to verify that the app
compiles successfully.

### Why is this needed?  
We need this workflow to enable
[Renovate](https://github.com/renovatebot/renovate) in this repository
and automatically update the mobile framework. Also serving as a way to
expose changes in the framework that could break apps consuming it.

### Implementation Details  
- The builds use the `eas` CLI in [local
mode](https://docs.expo.dev/build-reference/local-builds/) to simplify
the setup and ensure compatibility with EAS.
- We lose granular control over build steps (e.g., Fastlane, CocoaPods),
but this should be easier for builders
- This approach requires an Expo account, but a free account is
sufficient.
- The actual builds run on the runner.  

### Future Improvements  
- Add an additional step to verify that the app runs successfully after
compilation. (e.g. using Maestro).

### Performance Considerations  
We are using **custom runners** to speed up the workflow, as GitHub’s
default runners are too slow for our needs.

| Runner  | iOS (Total Time) | Android (Total Time) |
|---------|----------------|----------------|
| GitHub  | 29 mins        | 18 mins        |
| Custom  | 10 mins        | 11 mins        |